Adding big int check for parameters

+++ db/migrate/20190416143409_add_topic_id_to_assignment_questionnaires.rb
@@ -1,5 +1,5 @@
 class AddTopicIdToAssignmentQuestionnaires < ActiveRecord::Migration[4.2]
   def change
-    add_column :assignment_questionnaires, :topic_id, :integer  # might break
+    add_column :assignment_questionnaires, :topic_id, :bigint  # might break
   end
 end
